// src/routes/_private/ia/prompt/@interface/prompt.types.ts
import { z } from 'zod';

export const aiPromptSchema = z.object({
  question: z.string().min(1),
});

export type AIPromptData = z.infer<typeof aiPromptSchema>;

export interface ChatMessage {
  message: string;
  date: string;
  reply: boolean;
  type: 'text';
  sender: string;
  avatar?: string;
}